ingestion/requirements.txt
pandas==2.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
sqlalchemy==2.0.23